refactor: isolate swagger-ui-react DOM/CSS theming into swaggerTheme.ts
ApiDocumentation.tsx mixed page rendering logic with ~270 lines of
swagger-ui-react DOM-patching (enforceSwaggerA11yStyles, hard-coded
against its rendered class names) and CSS-in-JS overrides
(BASE_CSS/LIGHT_EXTRA/DARK_EXTRA). Both are fragile to a swagger-ui-react
version bump renaming its classes, with no compile-time signal, and were
untestable in isolation from the page component (which pulls in the
private @sethbacon/terraform-suite-ui package).

Move both into src/utils/swaggerTheme.ts with a comment noting the
verified swagger-ui-react version and the manual-QA steps a dependency
bump needs, plus a unit test exercising enforceSwaggerA11yStyles against
DOM fixtures and getSwaggerThemeCss's light/dark output -- this module
has no dependency on swagger-ui-react or the private package, so it can
run in this sandbox where the page-level test cannot.

No behavioral change.

Fixes #498 (item 31)
